+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.5.2] - 2020-05-18
+- Remove `adminCollectionSearch` use `searchCollections` instead
+- Add `getMyCollections` for convenience
+
 ## [10.5.1] - 2020-05-18
 - Rename `public` to `discoverable` (in collections)